training_center_admin_panel-main/requirements.txt
aioredis==1.3.1
alembic==1.5.7
amqp==5.0.6
anyio==3.1.0
appdirs==1.4.3
apturl==0.5.2
asgi-redis==1.4.3
asgiref==3.3.4
astroid==2.4.2
async-timeout==3.0.1
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
bcrypt==3.1.7
beat==0.3.0
beautifulsoup4==4.9.3
billiard==3.6.3.0
blinker==1.4
Brlapi==0.7.0
cairocffi==0.9.0
celery==5.0.5
celerybeat-redis==0.1.5
certifi==2019.11.28
cffi==1.14.5
channels==3.0.3
channels-redis==3.2.0
chardet==3.0.4
Click==7.0
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.7
cupshelpers==1.0
cycler==0.10.0
daphne==3.0.2
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.1.8
django-celery-results==2.0.1
django-channels==0.7.0
django-ckeditor==6.0.0
django-constance==2.8.0
django-js-asset==1.2.2
django-picklefield==3.0.1
django-rest-marshmallow==4.0.2
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
docopt==0.6.2
duplicity==0.8.12.0
entrypoints==0.3
falcon==2.0.0
fasteners==0.14.1
filelock==3.0.12
flower==0.9.7
future==0.18.2
greenlet==1.0.0
h11==0.12.0
h2==3.2.0
hiredis==2.0.0
hpack==3.0.0
http3==0.6.7
httpcore==0.13.4
httplib2==0.14.0
httpx==0.18.1
humanize==3.3.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==21.3.0
isort==5.7.0
Jinja2==2.11.2
keyring==18.0.1
kiwisolver==1.0.1
kombu==5.0.2
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.0
matplotlib==3.1.2
mccabe==0.6.1
monotonic==1.5
more-itertools==4.2.0
msgpack==1.0.2
msgpack-python==0.5.6
mysqlclient==2.0.3
netifaces==0.10.4
numpy==1.17.4
oauthlib==3.1.0
olefile==0.46
packaging==20.3
paramiko==2.6.0
pbr==5.5.1
pexpect==4.6.0
Pillow==8.1.1
pipreqs==0.4.10
playsound==1.2.2
ply==3.11
prometheus-client==0.8.0
prompt-toolkit==3.0.18
protobuf==3.6.1
psutil==5.5.1
py3dns==3.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycparser==2.20
pycups==1.9.73
Pygments==2.3.1
PyGObject==3.36.0
PyJWT==1.7.1
pylint==2.6.0
pymacaroons==0.13.0
PyMySQL==1.0.2
PyNaCl==1.3.0
pyOpenSSL==20.0.1
pyparsing==2.4.6
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.5
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-editor==1.0.4
python-memcached==1.59
pytz==2020.4
pyxdg==0.26
PyYAML==5.3.1
redis==3.2.0
reportlab==3.5.34
requests==2.25.1
requests-async==0.6.2
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rfc3986==1.5.0
SecretStorage==2.3.1
service-identity==18.1.0
simpleaudio==1.0.4
simplejson==3.16.0
six==1.15.0
sniffio==1.2.0
soupsieve==2.2.1
speedtest-cli==2.1.2
SQLAlchemy==1.4.2
sqlalchemy-migrate==0.13.0
sqlparse==0.4.1
starlette==0.14.2
starlette-request-id==0.3.6
stripe==2.56.0
systematic==4.8.7
systemd-python==234
Tempita==0.5.2
toml==0.10.2
tornado==6.1
Twisted==21.2.0
txaio==21.2.1
ubuntu-advantage-tools==27.0
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
urlopen==1.0.0
usb-creator==0.3.7
validate-email==1.3
vine==5.0.0
virtualenv==20.0.17
wadllib==1.3.3
wcwidth==0.2.5
wrapt==1.12.1
xcffib==0.8.1
xkit==0.0.0
yarg==0.1.9
zipp==1.0.0
zope.interface==5.4.0
